Move `RemoteFeedLoader` components to production
