+++ EssentialFeed/EssentialFeedTests/RemoteFeedLoaderTests.swift
@@ -6,25 +6,7 @@
 //
 
 import XCTest
-
-class RemoteFeedLoader {
-	
-	private let client: HTTPClient
-	private let url: URL
-	
-	init(url: URL, client: HTTPClient) {
-		self.url = url
-		self.client = client
-	}
-	
-	func load() {
-		client.get(from: url)
-	}
-}
-
-protocol HTTPClient {
-	func get(from url: URL)
-}
+import EssentialFeed
 
 class RemoteFeedLoaderTests: XCTestCase {
